# api-docs.yaml
openapi: 3.0.1
info:
  title: SROOM 프로젝트
  description: RESTful API
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /members/login:
    post:
      tags:
        - 로그인
      summary: 사용자 로그인
      description: 구글 ID 키를 사용하여 사용자를 인증합니다.
      operationId: login
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/GoogleIdKey'
        required: true
      responses:
        "200":
          description: 성공적으로 사용자를 인증하였습니다.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Login'
        "401":
          description: 인증에 실패하였습니다.
        "400":
          description: 요청 본문이 유효하지 않습니다.
  /lectures:
    get:
      tags:
        - 강의 검색
      summary: 강의 키워드 검색
      description: 키워드를 입력받아 유튜브 강의를 검색한다.
      operationId: getLecturesByKeyword
      parameters:
        - name: keyword
          in: query
          description: 검색할 키워드
          required: true
          schema:
            type: string
          example: 네트워크
        - name: limit
          in: query
          description: 결과의 최대 개수
          required: false
          schema:
            type: integer
            format: int32
            default: 10
          example: 5
        - name: nextPageToken
          in: query
          description: 다음 페이지 토큰
          required: false
          schema:
            type: string
          example: QAUQAA
        - name: prevPageToken
          in: query
          description: 이전 페이지 토큰
          required: false
          schema:
            type: string
          example: CAUQAA
      responses:
        "400":
          description: keyword가 입력되지 않았습니다.
        "200":
          description: 성공적으로 검색 결과를 반환하였습니다.
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/KeywordSearchRes'
  /lectures/{lectureId}:
    get:
      tags:
        - 강의 검색
      summary: 강의 상세 정보 조회
      description: 강의 ID를 이용하여 강의의 상세 정보를 조회한다.
      operationId: getLectureDetail
      parameters:
        - name: lectureId
          in: path
          description: 강의 ID
          required: true
          schema:
            type: string
          example: OEV8gMkCHXQ
        - name: is_playlist
          in: query
          description: 플레이리스트 여부
          required: true
          schema:
            type: boolean
          example: false
        - name: index_limit
          in: query
          description: 결과의 최대 개수
          required: false
          schema:
            type: integer
            format: int32
            default: 10
          example: 10
        - name: review_limit
          in: query
          description: 리뷰의 최대 개수
          required: false
          schema:
            type: integer
            format: int32
            default: 10
          example: 10
      responses:
        "404":
          description: 입력한 lectureId에 해당하는 강의가 없습니다.
        "200":
          description: 성공적으로 강의 상세 정보를 반환하였습니다.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PlaylistDetail'
                  - $ref: '#/components/schemas/VideoDetail'
        "400":
          description: 필수 파라미터인 'is_playlist'(boolean)가 누락되었습니다.
components:
  schemas:
    GoogleIdKey:
      type: object
      properties:
        credential:
          type: string
          description: 구글 ID 토큰
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6Ijk5YmNiMDY5MzQwYTNmMTQ3NDYyNzk0ZGZlZmE3NWU3OTk2MTM2MzQiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJuYmYiOjE2ODgyMTkyODAsImF1ZCI6IjEwODUyMjEwNDc5NDYtNHMxZ2YwZjlqdHQ3OTB2a2ZkcHZ1Z3FqbW9qMzdrdGsuYXBwcy5nb29nbGV1c2VyY29udGVudC5jb20iLCJzdWIiOiIxMDYwMzMxMTQ0MzU1MjY2MTA0ODciLCJlbWFpbCI6ImxqanVuMDcwMkBnbWFpbC5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiYXpwIjoiMTA4NTIyMTA0Nzk0Ni00czFnZjBmOWp0dDc5MHZrZmRwdnVncWptb2ozN2t0ay5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsIm5hbWUiOiLsnbTsooXspIAiLCJwaWN0dXJlIjoiaHR0cHM6Ly9saDMuZ29vZ2xldXNlcmNvbnRlbnQuY29tL2EvQUFjSFR0ZXQ5SnVoUUJQLWM1dEN0TXNLQmhRemdtZnNMdXpJVEhNS081bHNnN3FrPXM5Ni1jIiwiZ2l2ZW5fbmFtZSI6IuyiheykgCIsImZhbWlseV9uYW1lIjoi7J20IiwiaWF0IjoxNjg4MjE5NTgwLCJleHAiOjE2ODgyMjMxODAsImp0aSI6IjAxMTljNzBhZmQ5ODZkNzUxMzI3YjMxNzA3ZjVkNmVmNmMyM2JlZDgifQ.U1zcrnD9T_23bmT-8FQd_yY2PV9tnq2boIRZ9m37fVqeq25elAyBtghKG1p4mzbI1VoygbzdbyYihUkIVz8I21EDiiny9pO8ZBh7_pfKwId8UjeEQeJtrfidmYFgAPJ0dl4BoVZkISVHXok2XWPtWtzp3SHFKwePhW-SD0HAh8vDPJrkGCiQj5CupAL5nIs-GlIfusC7rx9A1hHaxu-CNyuEDlqrHh-MHlQLN4Wp42RfptOf68cIZZvKQ9D-Fx7Ag77HKOcikOFVk6Xx01mR0XtEaVsozPNRFudcrbcUMBdYYqrUjY_KN55MMA_ZPYeDbl67dNkyc6F3Ot8Wfq5yZA
      description: 구글 ID 토큰
    Login:
      type: object
      properties:
        accessToken:
          type: string
          description: 엑세스 토큰
          example: eyJhbGci...
        refreshToken:
          type: string
          description: 리프레시 토큰
          example: eyJhbGci...
        expireIn:
          type: integer
          description: 만료 시간(초)
          format: int64
          example: 1688398507
        memberName:
          type: string
          description: 멤버 이름
          example: user_978538
        bio:
          type: string
          description: 사용자 한줄소개
      description: 로그인 정보
    KeywordSearchRes:
      type: object
      properties:
        resultPerPage:
          type: integer
          description: 응답 개수
          format: int32
          example: 5
        nextPageToken:
          type: string
          description: 다음페이지 조회 토큰
          example: CAQQAA
        prevPageToken:
          type: string
          description: 이전페이지 조회 토큰
          example: QAQQAA
        lectures:
          type: array
          description: 조회된 강의 리스트
          items:
            $ref: '#/components/schemas/Lecture'
      description: 검색 결과 강의 리스트
    Lecture:
      type: object
      properties:
        lectureTitle:
          type: string
          description: 강의 제목
          example: 네트워크 기초(개정판)
        description:
          type: string
          description: 강의 설명
          example: OSI 7계층에서 각 계층의 다양한 프로토콜들을 통해서 배우는 네트워크 기초에 대한 강의입니다.
        lectureId:
          type: string
          description: 강의 ID
          example: PL0d8NnikouEWcF1jJueLdjRIC4HsUlULi
        rating:
          type: number
          description: 강의 평점
          format: double
          example: 4.3
        reviewCount:
          type: integer
          description: 후기 개수
          format: int32
          example: 44
        thumbnail:
          type: string
          description: 강의 썸네일
          example: https://i.ytimg.com/vi/Av9UFzl_wis/hqdefault.jpg
        isPlaylist:
          type: boolean
          description: 플레이리스트 여부
          example: true
      description: 검색된 개별 강의 정보
    Index:
      type: object
      properties:
        index:
          type: integer
          description: 목차 번호
          format: int32
          example: 1
        lectureId:
          type: string
          description: 강의 ID
          example: PL0d8NnikouEWcF1jJueLdjRIC4HsUlULi
        thumbnail:
          type: string
          description: 강의 썸네일
          example: https://i.ytimg.com/vi/Av9UFzl_wis/hqdefault.jpg
        lectureTitle:
          type: string
          description: 강의 제목
          example: 네트워크 기초(개정판)
      description: 재생목록 목차 정보
    PlaylistDetail:
      type: object
      properties:
        lectureId:
          type: string
          description: 재생목록 ID
          example: PL0d8NnikouEWcF1jJueLdjRIC4HsUlULi
        lectureTitle:
          type: string
          description: 재생목록 제목
          example: 네트워크 기초(개정판)
        channel:
          type: string
          description: 채널 이름
          example: Network Academy
        description:
          type: string
          description: 재생목록 설명
          example: OSI 7계층에서 각 계층의 다양한 프로토콜들을 통해서 배우는 네트워크 기초에 대한 강의입니다.
        publishedAt:
          type: string
          description: 재생목록 게시 날짜
          example: 2022-12-31T23:59:59Z
        lectureCount:
          type: integer
          description: 강의 개수
          format: int32
          example: 20
        rating:
          type: number
          description: 강의 평점
          format: double
          example: 4.3
        reviewCount:
          type: integer
          description: 후기 개수
          format: int32
          example: 44
        thumbnail:
          type: string
          description: 재생목록 썸네일
          example: https://i.ytimg.com/vi/Av9UFzl_wis/hqdefault.jpg
        indexNextPageToken:
          type: string
          description: 목차 다음 페이지 토큰
          example: CAoQAA
        indexes:
          type: array
          description: 강의 목차 정보
          items:
            $ref: '#/components/schemas/Index'
        reviews:
          type: array
          description: 강의 후기 요약 정보
          items:
            $ref: '#/components/schemas/ReviewBrief'
        isPlaylist:
          type: boolean
          description: 플레이리스트 여부
          example: true
      description: 재생목록 상세 정보
    ReviewBrief:
      type: object
      properties:
        index:
          type: integer
          format: int32
        reviewContent:
          type: string
        submittedRating:
          type: integer
          format: int32
      description: 강의 리뷰 목록
    VideoDetail:
      type: object
      properties:
        lectureId:
          type: string
          description: 강의 ID
          example: OEV8gMkCHXQ
        lectureTitle:
          type: string
          description: 강의 제목
          example: 네트워크 기초(개정판)
        channel:
          type: string
          description: 채널 이름
          example: OpenAI
        description:
          type: string
          description: 강의 설명
          example: OSI 7계층에서 각 계층의 다양한 프로토콜들을 통해서 배우는 네트워크 기초에 대한 강의입니다.
        viewCount:
          type: integer
          description: 조회 수
          format: int64
          example: 10234
        publishedAt:
          type: string
          description: 게시일
          example: 2023-06-28
        rating:
          type: number
          description: 강의 평점
          format: double
          example: 4.3
        reviewCount:
          type: integer
          description: 후기 개수
          format: int32
          example: 44
        thumbnail:
          type: string
          description: 강의 썸네일
          example: https://i.ytimg.com/vi/OEV8gMkCHXQ/maxresdefault.jpg
        reviews:
          type: array
          description: 강의 리뷰 목록
          items:
            $ref: '#/components/schemas/ReviewBrief'
        isPlaylist:
          type: boolean
          description: 플레이리스트 여부
          example: false
      description: 동영상 상세 정보